move and clean up files
